fixed something time for tests

+++ utils/utils.py
@@ -1,8 +1,8 @@
 import re
 
 
-def findTempAndDegree(message: str) -> tuple[int, str] | None:
+def findTempAndDegree(message: str) -> tuple[float, str] | None:
     v = re.search(r'-?(\d+|\d*[.,]\d+|\d+.\d*)\s*°?\s*(C|F|c|f)\b', message)
     if v == None:
         return None
-    return (int(v.groups()[0]), v.groups()[1])
+    return (float(v.groups()[0]), v.groups()[1])
